# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmPay API",
        "description": "API de gestion bancaire avec authentification Passport",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://khadidiatou-fall-api-laravel-0luq.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Informations de l'utilisateur connecté",
                "description": "Récupérer les informations de l'utilisateur connecté avec son compte et ses transactions",
                "operationId": "1e5f382ba4fa0960d233875cab75a361",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page pour les transactions",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre de transactions par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer les transactions par type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "depot",
                                "retrait",
                                "transfert",
                                "payement"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer les transactions par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en_cours",
                                "validee",
                                "rejete",
                                "annule"
                            ]
                        }
                    },
                    {
                        "name": "date_debut",
                        "in": "query",
                        "description": "Date de début (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_fin",
                        "in": "query",
                        "description": "Date de fin (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Trier par champ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "date_transaction",
                            "enum": [
                                "date_transaction",
                                "montant",
                                "type"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations de l'utilisateur récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        },
                                                        "nom": {
                                                            "type": "string"
                                                        },
                                                        "prenom": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        },
                                                        "telephone": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "compte": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        },
                                                        "numeroCompte": {
                                                            "type": "string"
                                                        },
                                                        "solde": {
                                                            "type": "number",
                                                            "format": "float"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "transactions": {
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/Transaction"
                                                            }
                                                        },
                                                        "pagination": {
                                                            "$ref": "#/components/schemas/TransactionPagination"
                                                        },
                                                        "links": {
                                                            "$ref": "#/components/schemas/TransactionLinks"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Utilisateur ou compte non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TransactionPagination": {
                "title": "Pagination des Transactions",
                "description": "Informations de pagination pour les transactions",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "TransactionLinks": {
                "title": "Liens de navigation des Transactions",
                "description": "Liens de navigation pour les transactions",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://api.banque.example.com/api/v1/transactions?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://api.banque.example.com/api/v1/transactions?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://api.banque.example.com/api/v1/transactions?page=2"
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "title": "Transaction",
                "description": "Objet représentant une transaction bancaire",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "numeroTransaction": {
                        "type": "string",
                        "example": "TXN-20251102-ABCD"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "depot",
                            "retrait",
                            "transfert",
                            "payement"
                        ],
                        "example": "depot"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "description": {
                        "type": "string",
                        "example": "Dépôt d'espèces",
                        "nullable": true
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "en_cours",
                            "validee",
                            "rejete",
                            "annule"
                        ],
                        "example": "validee"
                    },
                    "dateTransaction": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-02T10:30:00Z"
                    },
                    "compteSource": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "numeroCompte": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "compteDestination": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "numeroCompte": {
                                "type": "string"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-11-02T10:30:00Z"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}